# doc/api.yaml
openapi: 3.0.3
info:
  title: Система бронирования столов в офисе
  version: '1.0'
...
paths:
...
  /offices/{officeId}:   
    delete:
      tags:
        - office
      summary: Удаление данных о офисе
      operationId: deleteOffices
      description: Удаление данных о офисе по id
      parameters:
        - name: officeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Офис не найден
    put:
      tags:
        - office
      summary: Обновление данных о офисе
      description: Обновиление данных о офисе по id
      operationId: updateOffice
      parameters:
        - name: officeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Данные о офисе
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Office'
        required: true
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 12
        '404':
          description: Офис не найден
    get:
      tags:
        - office
      summary: Получение данных о офисе
      description: Получение данных о офисе по id
      operationId: getOfficesById
      parameters:
        - name: officeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Office'
        '404':
          description: Офис не найден
...
components:
  schemas:
    Office:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          description: Название офиса
          example: r-901
        location:
          type: json
...